backend/requirements.txt
# GitLab Handbook AI — Python dependencies

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# RAG core
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4

# LLM
anthropic==0.40.0
groq==0.13.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
